Improved the feedback management component with data visualization

+++ app/api/feedback/[id]/route.ts
@@ -1,11 +1,17 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/prisma/client";
 import { feedbackSchema } from "@/validations/schema.validation";
+import { ObjectId } from "mongodb";
 
 export async function GET(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  // Check if the id is a valid ObjectId
+  if (!ObjectId.isValid(params.id)) {
+    return NextResponse.json({ error: "Invalid feedback ID" }, { status: 400 });
+  }
+
   const feedback = await prisma.feedback.findUnique({
     where: { id: params.id },
     include: { replies: true },
@@ -22,6 +28,11 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  // Check if the id is a valid ObjectId
+  if (!ObjectId.isValid(params.id)) {
+    return NextResponse.json({ error: "Invalid feedback ID" }, { status: 400 });
+  }
+
   const body = await request.json();
 
   const validation = feedbackSchema.safeParse(body);
@@ -56,6 +67,11 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  // Check if the id is a valid ObjectId
+  if (!ObjectId.isValid(params.id)) {
+    return NextResponse.json({ error: "Invalid feedback ID" }, { status: 400 });
+  }
+
   const feedback = await prisma.feedback.delete({
     where: { id: params.id },
   });
